+++ src/main/java/com/example/simple/mall/api/service/impl/UserServiceImpl.java
@@ -87,7 +87,7 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, UserEntity> impleme
      * @since 2025/05/24@return @return {@code Boolean }@return @return {@code Boolean }
      */
     @Override
-    public Boolean judgeUserIfNull(Integer userId) {
+    public Boolean judgeUserIfNull(Long userId) {
         QueryWrapper<UserEntity> userWrapper = new QueryWrapper<>();
         userWrapper.eq("id", userId);
         userWrapper.eq("status", UserStatusEnum.FRIEND.getCode());
@@ -103,7 +103,7 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, UserEntity> impleme
     @Override
     @Transactional(rollbackFor = Exception.class)
     public void updateUser(UserDTO userDto) {
-        Integer id = userDto.getId();
+        Long id = userDto.getId();
         if (ObjectUtil.isNotEmpty(id)) {
             UserEntity userOld = userMapper.selectById(id);
             if (ObjectUtil.isEmpty(userOld)) {
